timetable add

+++ database/migrations/2026_04_25_095635_create_assign_teacher_to_classes_table.php
@@ -11,15 +11,16 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('assign_teacher_to_classes', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('class_id')->constrained()->onDelete('cascade');
-            $table->foreignId('subject_id')->constrained()->onDelete('cascade');
-            $table->foreignId('teacher_id')->constrained()->onDelete('cascade');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('assign_teacher_to_classes')) {
+            Schema::create('assign_teacher_to_classes', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('class_id')->constrained()->onDelete('cascade');
+                $table->foreignId('subject_id')->constrained()->onDelete('cascade');
+                $table->foreignId('teacher_id')->constrained('users')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
-
     /**
      * Reverse the migrations.
      */
